_model_line --> _model_repr & Model2D plot outline
