+++ hyperspy/models/model1D.py
@@ -19,7 +19,6 @@
 import copy
 
 import numpy as np
-from contextlib import contextmanager
 
 from hyperspy.model import BaseModel, ModelComponents, ModelSpecialSlicers
 import hyperspy.drawing.spectrum
@@ -169,6 +168,7 @@ class Model1D(BaseModel):
         self._plot_components = False
         self._suspend_update = False
         self._model_line = None
+        self._model_repr = self._model_line
         self._adjust_position_all = None
         self.axis = self.axes_manager.signal_axes[0]
         self.axes_manager.events.indices_changed.connect(
@@ -274,86 +274,6 @@ class Model1D(BaseModel):
         self._disconnect_parameters2update_plot(things)
 
     remove.__doc__ = BaseModel.remove.__doc__
-
-    def _connect_parameters2update_plot(self, components):
-        if self._plot_active is False:
-            return
-        for i, component in enumerate(components):
-            component.events.active_changed.connect(
-                self._model_line.update, [])
-            for parameter in component.parameters:
-                parameter.events.value_changed.connect(
-                    self._model_line.update, [])
-        if self._plot_components is True:
-            self._connect_component_lines()
-
-    def _disconnect_parameters2update_plot(self, components):
-        if self._model_line is None:
-            return
-        for component in components:
-            component.events.active_changed.disconnect(self._model_line.update)
-            for parameter in component.parameters:
-                parameter.events.value_changed.disconnect(
-                    self._model_line.update)
-        if self._plot_components is True:
-            self._disconnect_component_lines()
-
-    def update_plot(self, *args, **kwargs):
-        """Update model plot.
-
-        The updating can be suspended using `suspend_update`.
-
-        See Also
-        --------
-        suspend_update
-
-        """
-        if self._plot_active is True and self._suspend_update is False:
-            try:
-                self._update_model_line()
-                for component in [component for component in self if
-                                  component.active is True]:
-                    self._update_component_line(component)
-            except:
-                self._disconnect_parameters2update_plot(components=self)
-
-    @contextmanager
-    def suspend_update(self, update_on_resume=True):
-        """Prevents plot from updating until 'with' clause completes.
-
-        See Also
-        --------
-        update_plot
-        """
-
-        es = EventSupressor()
-        es.add(self.axes_manager.events.indices_changed)
-        if self._model_line:
-            f = self._model_line.update
-            for c in self:
-                es.add(c.events, f)
-                for p in c.parameters:
-                    es.add(p.events, f)
-        for c in self:
-            if hasattr(c, '_model_plot_line'):
-                f = c._model_plot_line.update
-                es.add(c.events, f)
-                for p in c.parameters:
-                    es.add(p.events, f)
-
-        old = self._suspend_update
-        self._suspend_update = True
-        with es.suppress():
-            yield
-        self._suspend_update = old
-
-        if update_on_resume is True:
-            self.update_plot()
-
-    def _update_model_line(self):
-        if (self._plot_active is True and
-                self._model_line is not None):
-            self._model_line.update()
 
     def __call__(self, non_convolved=False, onlyactive=False):
         """Returns the corresponding model for the current coordinates
@@ -623,7 +543,7 @@ class Model1D(BaseModel):
         on_figure_window_close(_plot.signal_plot.figure,
                                self._close_plot)
 
-        self._model_line = l2
+        self._model_repr = l2
         self._plot = self.spectrum._plot
         self._connect_parameters2update_plot(self)
         if plot_components is True:
@@ -680,12 +600,6 @@ class Model1D(BaseModel):
             component._model_plot_line.close()
             del component._model_plot_line
         self._plot_components = False
-
-    def _close_plot(self):
-        if self._plot_components is True:
-            self.disable_plot_components()
-        self._disconnect_parameters2update_plot(components=self)
-        self._model_line = None
 
     def enable_plot_components(self):
         if self._plot is None or self._plot_components:
